+++ test/integration/client/batch_delete_test.dart
@@ -7,6 +7,7 @@ import 'package:nostr_mail/src/storage/models/email_record.dart';
 import 'package:test/test.dart';
 
 import '../../helpers/test_user.dart';
+import '../../helpers/wait_for_broadcasts.dart';
 import '../../mocks/mock_relay.dart';
 
 void main() {
@@ -72,7 +73,7 @@ void main() {
       );
 
       await user.client.delete(['email-1', 'email-2']);
-      await user.client.flushBroadcasts();
+      await waitForBroadcasts(user.client.broadcastQueue);
 
       expect(await user.client.getEmail('email-1'), isNull);
       expect(await user.client.getEmail('email-2'), isNull);
@@ -158,7 +159,7 @@ void main() {
       );
 
       await user.client.delete(['email-1']);
-      await user.client.flushBroadcasts();
+      await waitForBroadcasts(user.client.broadcastQueue);
 
       final deletionEvents = await user.ndk.requests
           .query(